+++ lib/store/event/monetization_settings.dart
@@ -18,6 +18,7 @@
 import '/domain/model/chat.dart';
 import '/domain/model/monetization_settings.dart';
 import '/domain/model/precise_date_time/precise_date_time.dart';
+import '/domain/model/user.dart';
 import '/store/model/monetization_settings.dart';
 
 /// Possible kinds of a [MonetizationSettingsEvent].
@@ -107,7 +108,11 @@ class MonetizationSettingsEventsVersioned {
 
 /// Events happening in a [Chat].
 abstract class MonetizationSettingsEvent {
-  const MonetizationSettingsEvent(this.monetizationSettings, this.at);
+  const MonetizationSettingsEvent(
+    this.monetizationSettings,
+    this.userId,
+    this.at,
+  );
 
   /// State of the [MonetizationSettings] after this [MonetizationSettingsEvent]
   /// being applied.
@@ -115,6 +120,9 @@ abstract class MonetizationSettingsEvent {
 
   /// [PreciseDateTime] when this [MonetizationSettingsEvent] happened.
   final PreciseDateTime at;
+
+  /// [UserId] the [MonetizationSettings] are specified individually for.
+  final UserId? userId;
 
   /// Returns [MonetizationSettingsEventKind] of this [MonetizationSettingsEvent].
   MonetizationSettingsEventKind get kind;
@@ -125,6 +133,7 @@ class EventMonetizationSettingsDonationDeleted
     extends MonetizationSettingsEvent {
   const EventMonetizationSettingsDonationDeleted(
     super.monetizationSettings,
+    super.userId,
     super.at,
   );
 
@@ -138,6 +147,7 @@ class EventMonetizationSettingsDonationMinPriceUpdated
     extends MonetizationSettingsEvent {
   const EventMonetizationSettingsDonationMinPriceUpdated(
     super.monetizationSettings,
+    super.userId,
     super.at,
   );
 
@@ -151,6 +161,7 @@ class EventMonetizationSettingsDonationToggled
     extends MonetizationSettingsEvent {
   const EventMonetizationSettingsDonationToggled(
     super.monetizationSettings,
+    super.userId,
     super.at,
   );